requirements.txt
Django==3.2.12
django-environ==0.8.1
psycopg2-binary>=2.8
djangorestframework==3.13.1
django-filter>= 2.0.0